requirements.txt
Pillow==5.2.0
Django==2.1.15
django-ckeditor==5.9.0
django-js-asset==1.2.2